inline config handling into project
…as well as all dependent code
