+++ pacesetter-cli/blueprint/config/Cargo.toml.liquid
@@ -5,5 +5,8 @@ edition = "2021"
 publish = false
 
 [dependencies]
+anyhow = "1"
+dotenvy = "0.15"
+figment = { version = "0.10", features = ["toml", "env"] }
 pacesetter = { workspace = true }
 serde = { version = "1.0", features = ["derive"] }
