+++ pacesetter-cli/blueprint/web/Cargo.toml.liquid
@@ -25,6 +25,10 @@ uuid = { version = "1.6", features = ["serde"] }
 
 [dev-dependencies]
 fake = "2.9"
+rand = "0.8"
+regex = "1.10"
+# TODO: remove this later!
+sqlx = { version = "0.7", features = [ "runtime-tokio", "tls-rustls", "postgres", "migrate" ] }
 serde_json = "1.0"
 tower = { version = "0.4", features = ["util"] }
 hyper = { version = "1.0", features = ["full"] }
