+++ pacesetter/src/cli/mod.rs
@@ -1,4 +1,1 @@
-pub(crate) mod util;
-
-pub mod db;
 pub mod generate;
